+++ src/sockets/dashboard/__init__.py
@@ -1,0 +1,23 @@
+from .core import *
+from .cache_utils import (
+    SelectiveCache,
+    selective_cache,
+    _hash_cache,
+    _correlation_cache,
+    _success_cache,
+    _classic_stats_cache,
+    _new_stats_cache,
+    _team_process_cache,
+    _make_cache_key,
+)
+from .core import (
+    _process_single_team,
+    _process_single_team_optimized,
+    _calculate_team_statistics,
+    _calculate_success_statistics,
+    _calculate_team_statistics_from_data,
+    _calculate_success_statistics_from_data,
+    _compute_team_hashes_optimized,
+    _compute_correlation_matrix_optimized,
+    _compute_success_metrics_optimized,
+)
